Create README.md
Hi!
When I first saw the repository i didint really got what it really does or how to achieve it, so that i had to go deep in the code just to see the code.
I do believe that was not supposed to. So that I simple did added a README file with brief description and an how to use!